backend/requirements.txt
fastapi
uvicorn[standard]
pydantic

sqlalchemy
passlib[bcrypt]
bcrypt==3.2.2

pandas
numpy
scikit-learn
pyarrow

google-generativeai
httpx
parsel

rapidfuzz
deep-translator
tqdm
python-dotenv